rag-backend/requirements.txt
# Core Dependencies
python-docx==1.0.1
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings
pydantic[email]>=2.0
tenacity==8.2.3  # For retry logic

# Web Framework
fastapi[standard]==0.115.12
uvicorn[standard]==0.24.0  # ASGI server for FastAPI

# Database and ORM
sqlalchemy==2.0.23
asyncpg==0.29.0
pgvector
redis==5.0.1  # For caching

# Document Parsing
PyMuPDF==1.23.8  # for PDF text extraction

# RAG & Embedding Stack
transformers==4.36.2
torch==2.1.2+cpu  # CPU-only version to avoid NVIDIA libraries
--extra-index-url https://download.pytorch.org/whl/cpu
openai==1.3.5  # for answer generation

# Machine Learning
numpy==1.26.2

# Environment and Configuration
python-dotenv==1.0.0

# Utilities
python-magic  # for file type detection
tqdm  # for progress bars
psutil==5.9.8  # for memory monitoring

# Testing Dependencies
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
pytest-benchmark==4.0.0
aioresponses==0.7.5
pytest-mock==3.12.0